chore: seperated project components for CE
