+++ web/ee/components/projects/header.tsx
@@ -1,0 +1,1 @@
+export * from "ce/components/projects/header";
