+++ web/ee/components/projects/mobile-header.tsx
@@ -1,0 +1,1 @@
+export * from "ce/components/projects/mobile-header";
