+++ web/ee/components/projects/page.tsx
@@ -1,0 +1,1 @@
+export * from "ce/components/projects/page";
